Allow multi customer delimiters and last reafctoring

+++ Benchmark.Calculator.Application/Contracts/IValidationStrategy.cs
@@ -2,6 +2,6 @@
 {
     public interface IValidationStrategy
     {
-        bool Validate(int input);
+        bool Validate(long input);
     }
 }
