+++ Benchmark.Calculator.Application/Services/Validation/NegativeNumberValidation.cs
@@ -4,7 +4,7 @@ namespace Benchmark.Calculator.Application.Services.Validation
 {
     public class NegativeNumberValidation : IValidationStrategy
     {
-        public bool Validate(int input)
+        public bool Validate(long input)
         {
             if (input < 0)
                 return false;
